Refactored AppGenerator and PluginNewGenerator to inherit from AppBase.
